Create product dto

+++ src/migrations/1593778950300-Products.ts
@@ -63,7 +63,7 @@ export class Products1593778950300 implements MigrationInterface {
           },
           {
             name: 'price',
-            type: 'float',
+            type: 'numeric',
             isNullable: false,
           },
           {
